datatable and authorize role

+++ app/Http/Controllers/Admin/PermissionsController.php
@@ -2,11 +2,13 @@
 
 namespace App\Http\Controllers\Admin;
 
+use Gate;
 use App\Models\Permission;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\StorePermissionRequest;
 use App\Http\Requests\UpdatePermissionRequest;
+use Symfony\Component\HttpFoundation\Response;
 
 class PermissionsController extends Controller
 {
@@ -17,6 +19,8 @@ class PermissionsController extends Controller
      */
     public function index()
     {
+        abort_if(Gate::denies('permission_access'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+
         $permissions = Permission::paginate(10);
         $firstItem = $permissions->firstItem();
         return view('admin.permissions.index',[
@@ -32,6 +36,8 @@ class PermissionsController extends Controller
      */
     public function create()
     {
+        abort_if(Gate::denies('permission_create'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+
         return view('admin.permissions.create');
     }
 
@@ -56,6 +62,8 @@ class PermissionsController extends Controller
      */
     public function show($id)
     {
+        abort_if(Gate::denies('permission_show'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+
         $permission = Permission::find($id);
         return view('admin.permissions.show',[
             'permission' => $permission
@@ -70,6 +78,8 @@ class PermissionsController extends Controller
      */
     public function edit($id)
     {
+        abort_if(Gate::denies('permission_edit'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+
         $permission = Permission::find($id);
         return view('admin.permissions.edit',[
             'permission' => $permission
@@ -98,6 +108,8 @@ class PermissionsController extends Controller
      */
     public function destroy(Permission $permission)
     {
+        abort_if(Gate::denies('permission_delete'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+
         $permission->delete();
         return redirect()->route('admin.permissions.index');
     }
